requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
googlemaps==4.6.0
idna==3.4
Pillow==9.2.0
requests==2.28.1
urllib3==1.26.12
